Move DumpIrTreeWithDescriptorsVisitor to common package
